+++ Fastly/CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add connector for Fastly Audit logs
+- Add connector for Fastly WAF Audit logs
 
 ## 2024-02-13 - 0.1.2